requirements.txt
crewai>=0.11.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
prophet>=1.1.4
xgboost>=2.0.0
google-generativeai>=0.3.0
python-telegram-bot>=20.0
apscheduler>=3.10.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0 